(PC-20410) refacto(datetimeHelpers): move datetime helpers closer to where they are used
