+++ src/libs/algolia/fetchAlgolia/buildAlgoliaParameters/helpers/datetime/__tests__/date.test.ts
@@ -1,4 +1,4 @@
-import { Range } from '../../../../libs/typesUtils/typeHelpers'
+import { Range } from '../../../../../../typesUtils/typeHelpers'
 import { DATE } from '../date'
 
 describe('getDatesFromTimeRangeAndDate', () => {
