+++ src/libs/algolia/fetchAlgolia/buildAlgoliaParameters/helpers/datetime/__tests__/time.test.ts
@@ -1,6 +1,6 @@
 import mockdate from 'mockdate'
 
-import { Range } from '../../../../libs/typesUtils/typeHelpers'
+import { Range } from '../../../../../../typesUtils/typeHelpers'
 import { computeTimeRangeFromHoursToSeconds, TIMESTAMP } from '../time'
 
 describe('Timezones', () => {
